Refactor account management to use MongoDB schema and update related tests
